+++ modules/documentation/docs/changelog.md
@@ -2,6 +2,26 @@
 
 ## Next Release
 
+## 0.2.1
+
+Non-beta release
+
+## 0.2.1-beta.24
+
+- \[router\] Fix autorebalancing for multiple rebalance cycles
+
+## 0.2.1-beta.23
+
+- \[router\] Fix autorebalancing approval tx schema
+
+## 0.2.1-beta.22
+
+- \[router\] Fix autorebalancing
+
+## 0.2.1-beta.21
+
+- \[router\] Potential double collateral edge case
+
 ## 0.2.1-beta.20
 
 - \[node\] Router config for cluster